No more terraform, ansible does everything
Sorry TF, this is like a recipe, so the declarative approach doesn't
work.

Everything works fine btw, except for the Instance Template doesn't
accept the (obsoleted) sshKeys metadata, so we can't log in to the
instantiated webservers.

Looks like we'll have to do it the clean (and official) way and
add those public keys on project-level. (Or enable OSLogin and
add IAM accounts, if finer access control is needed.)